+++ lib/modules/core/core_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: core
-// Updated at: 2026-06-03 13:39:40
-// Commit: chore(core): update dependencies and cleanup unused imports in core
+// Updated at: 2026-06-03 15:45:54
+// Commit: perf(core): apply RepaintBoundary on animated core widgets
 
 class coreComponentState {
-  final String commitRef = '1698';
-  final String lastMessage = 'chore(core): update dependencies and cleanup unused imports in core';
-  final DateTime timestamp = DateTime.parse('2026-06-03T13:39:40+05:00');
+  final String commitRef = '1708';
+  final String lastMessage = 'perf(core): apply RepaintBoundary on animated core widgets';
+  final DateTime timestamp = DateTime.parse('2026-06-03T15:45:54+05:00');
 }